# systemd: speedup monitor command 2x