mainly by reusing dbug property accessor (not 100% sure why it helps but it works)

before
```
$ hyperfine 'dron monitor --once'
Benchmark 1: dron monitor --once
  Time (mean ± σ):      1.993 s ±  0.066 s    [User: 0.976 s, System: 0.113 s]
  Range (min … max):    1.926 s …  2.118 s    10 runs
```

after
```
$ hyperfine 'dron monitor --once'
Benchmark 1: dron monitor --once
  Time (mean ± σ):     913.7 ms ±  71.9 ms    [User: 471.3 ms, System: 50.1 ms]
  Range (min … max):   850.3 ms … 1077.6 ms    10 runs
```